feat:캐시 관련 테스트코드 작성

+++ src/main/java/com/sparta/sportify/dto/cash/response/CashResponseDto.java
@@ -13,7 +13,7 @@ public class CashResponseDto {
 	private Long balance;
 	private LocalDateTime createAt;
 	private CashType type;
-	private Integer ChargeAmount;
+	private Long ChargeAmount;
 
 	public CashResponseDto(CashLog cashLog) {
 		this.userId = cashLog.getUser().getId();
